Remove Chrono include and utils.h->commons.h
