+++ include/seahowl/aero/aerodyn_adapter.h
@@ -7,7 +7,7 @@
 #include <seahowl/elasto/blade_elasto.h>
 #include <seahowl/elasto/rotor_elasto.h>
 #include <seahowl/elasto/elasto.h>
-#include <seahowl/utils.h>
+#include <seahowl/commons.h>
 
 #include <chrono/physics/ChBody.h>
 #include <chrono/core/ChMatrix33.h>
