+++ include/seahowl/aero/bemt.h
@@ -3,7 +3,7 @@
 #include <seahowl/aero/airfoil.h>
 #include <seahowl/aero/blade_aero.h>
 #include <seahowl/aero/tower_aero.h>
-#include <seahowl/utils.h>
+#include <seahowl/commons.h>
 
 namespace seahowl {
 namespace aero {
