+++ include/seahowl/aero/blade_aero.h
@@ -3,7 +3,7 @@
 #include <seahowl/aero/reference_point_aero.h>
 #include <seahowl/aero/wind_models.h>
 #include <seahowl/core/utils.h>
-#include <seahowl/utils.h>
+#include <seahowl/commons.h>
 
 namespace seahowl {
