+++ include/seahowl/aero/reference_point_aero.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <seahowl/core/reference_point.h>
-#include <seahowl/utils.h>
+#include <seahowl/commons.h>
 
 namespace seahowl {
 namespace aero {
