+++ include/seahowl/aero/wind_models.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "seahowl/utils.h"
+#include "seahowl/commons.h"
 
 namespace seahowl {
 namespace aero {
