+++ include/seahowl/elasto/reference_point_elasto.h
@@ -2,7 +2,7 @@
 
 #include <seahowl/core/reference_point.h>
 
-#include <seahowl/utils.h>
+#include <seahowl/commons.h>
 
 namespace seahowl {
 namespace elasto {
