+++ include/seahowl/elasto/rotor_elasto.h
@@ -5,7 +5,7 @@
 #include <vector>
 #include <memory>
 
-#include <seahowl/utils.h>
+#include <seahowl/commons.h>
 
 namespace seahowl {
 namespace elasto {
